refactor: view패키지에 남은 파일을 feature와 extension으로 이동
