+++ app/src/main/java/woowacourse/movie/extension/ImageResource.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.model
+package woowacourse.movie.extension
 
 import android.os.Parcelable
 import android.widget.ImageView
