+++ app/src/main/java/woowacourse/movie/feature/movieSelect/adapter/ScreeningData.kt
@@ -2,9 +2,9 @@ package woowacourse.movie.feature.movieSelect.adapter
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import woowacourse.movie.extension.ImageResource
 import woowacourse.movie.model.movieSelect.screening.Movie
 import woowacourse.movie.model.movieSelect.screening.Screening
-import woowacourse.movie.view.model.ImageResource
 import java.time.LocalDate
 
 @Parcelize
