+++ app/src/main/java/woowacourse/movie/feature/reservation/ReservationContract.kt
@@ -1,9 +1,9 @@
 package woowacourse.movie.feature.reservation
 
 import woowacourse.movie.feature.movieSelect.adapter.ScreeningData
+import woowacourse.movie.feature.ticket.TicketData
 import woowacourse.movie.model.movieSelect.screening.Screening
 import woowacourse.movie.model.ticket.TicketCount
-import woowacourse.movie.view.model.TicketData
 import java.time.LocalDate
 
 interface ReservationContract {
