+++ app/src/main/java/woowacourse/movie/feature/seatSelect/SeatIndexData.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.model
+package woowacourse.movie.feature.seatSelect
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
