+++ app/src/main/java/woowacourse/movie/feature/ticket/TicketContract.kt
@@ -1,7 +1,6 @@
 package woowacourse.movie.feature.ticket
 
 import woowacourse.movie.model.ticket.Ticket
-import woowacourse.movie.view.model.TicketData
 
 interface TicketContract {
     interface TicketView {
