+++ app/src/main/java/woowacourse/movie/feature/ticket/TicketData.kt
@@ -1,8 +1,9 @@
-package woowacourse.movie.view.model
+package woowacourse.movie.feature.ticket
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 import woowacourse.movie.feature.movieSelect.adapter.ScreeningData
+import woowacourse.movie.feature.seatSelect.SeatsData
 import java.time.LocalDateTime
 
 @Parcelize
@@ -10,7 +11,7 @@ data class TicketData(
     val screeningData: ScreeningData,
     val showtime: LocalDateTime,
     val ticketCount: Int,
-    val seatsData: SeatsData = SeatsData.getEmptySeatsData(),
+    val seatsData: SeatsData = SeatsData.Companion.getEmptySeatsData(),
 ) : Parcelable {
     fun seatsAddedTicketData(seatsData: SeatsData): TicketData = TicketData(this.screeningData, this.showtime, this.ticketCount, seatsData)
 }
